+++ tools/apidiff/report/packages.go
@@ -9,7 +9,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/Azure/azure-sdk-for-go/tools/apidiff/markdown"
+	"github.com/Azure/azure-sdk-for-go/tools/common/markdown"
 )
 
 const apiDirSuffix = "api"
